Added events added fake data

+++ Kancelarele/src/main/java/com/kancelarele/model/Case.java
@@ -18,7 +18,7 @@ public class Case {
     
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private int id;
+    private Long id;
     
     private String signature;
     
@@ -28,11 +28,11 @@ public class Case {
     @OneToMany(cascade = CascadeType.ALL, mappedBy = "eventCase")
     private List <Event> events;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -59,6 +59,8 @@ public class Case {
     public void setEvents(List<Event> events) {
         this.events = events;
     }
+
+   
 
     @Override
     public String toString() {
